task 1 completed

+++ src/main/java/com/aritoncosmin/ProiectSpringJava/repository/ContinentDataRepository.java
@@ -4,8 +4,8 @@ import com.aritoncosmin.ProiectSpringJava.model.Continent;
 import com.aritoncosmin.ProiectSpringJava.model.ContinentData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContinentDataRepository extends JpaRepository<ContinentData, Integer> {
+public interface ContinentDataRepository extends JpaRepository<ContinentData, Long> {
 
-    ContinentData findContinentDataByYearAndContinent(Integer year, Continent continent);
+    ContinentData findContinentDataByYearAndContinent(Long year, Continent continent);
 
 }
