+++ src/main/java/com/aritoncosmin/ProiectSpringJava/repository/ContinentRepository.java
@@ -3,7 +3,7 @@ package com.aritoncosmin.ProiectSpringJava.repository;
 import com.aritoncosmin.ProiectSpringJava.model.Continent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContinentRepository extends JpaRepository<Continent, Integer> {
+public interface ContinentRepository extends JpaRepository<Continent, Long> {
 
     Continent findContinentByName(String name);
