+++ src/main/java/com/aritoncosmin/ProiectSpringJava/repository/CountryDataRepository.java
@@ -4,7 +4,7 @@ import com.aritoncosmin.ProiectSpringJava.model.Country;
 import com.aritoncosmin.ProiectSpringJava.model.CountryData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryDataRepository extends JpaRepository<CountryData, Integer> {
-    CountryData findCountryDataByYearAndCountry(Integer year, Country country);
+public interface CountryDataRepository extends JpaRepository<CountryData, Long> {
+    CountryData findCountryDataByYearAndCountry(Long year, Country country);
 
 }
